Fix: views 함수명 통일 및 url 경로 수정

+++ free_board/views.py
@@ -12,14 +12,14 @@ from .serializers import (
 
 # 전체 게시물 조회
 @api_view(['GET'])
-def list_articles(request):
+def free_board_list(request):
     articles = FreeBoardArticle.objects.all()
     serializer = FreeBoardArticleListSerializer(articles, many=True)
     return Response(serializer.data)
 
 # 게시글 작성
 @api_view(['POST'])
-def create_article(request):
+def create_free_board_article(request):
     serializer = FreeBoardArticleSerializer(data=request.data)
     if serializer.is_valid():
         serializer.save(userId=request.user)
@@ -28,31 +28,30 @@ def create_article(request):
 
 # 게시글 상세 정보 조회
 @api_view(['GET'])
-def article_detail(request, article_pk):
+def free_board_article_detail(request, article_pk):
     article = get_object_or_404(FreeBoardArticle, pk=article_pk)
     serializer = FreeBoardArticleSerializer(article)
     return Response(serializer.data)
 
-# 게시글 수정
-@api_view(['PUT'])
-def update_article(request, article_pk):
-    article = get_object_or_404(FreeBoardArticle, pk=article_pk)
-    serializer = FreeBoardArticleSerializer(article, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-# 게시글 삭제
-@api_view(['DELETE'])
-def delete_article(request, article_pk):
+@api_view(['PUT', 'DELETE'])
+def edit_free_board_article(request, article_pk):
     article = get_object_or_404(FreeBoardArticle, pk=article_pk)
-    article.delete()
+    # 게시글 수정
+    if request.method == 'PUT':
+        serializer = FreeBoardArticleSerializer(article, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    # 게시글 삭제
+    else:
+        article.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
 
 # 댓글 작성
 @api_view(['POST'])
-def create_comment(request, article_pk):
+def create_free_board_comment(request, article_pk):
     article = get_object_or_404(FreeBoardArticle, pk=article_pk)
     serializer = FreeBoardCommentSerializer(data=request.data)
     if serializer.is_valid():
@@ -60,19 +59,30 @@ def create_comment(request, article_pk):
         return Response(serializer.data, status=status.HTTP_201_CREATED)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-# 댓글 수정
-@api_view(['PUT'])
-def update_comment(request, article_pk, comment_pk):
-    comment = get_object_or_404(FreeBoardComment, pk=comment_pk, article_id=article_pk)
-    serializer = FreeBoardCommentSerializer(comment, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-# 댓글 삭제
-@api_view(['DELETE'])
-def delete_comment(request, article_pk, comment_pk):
+@api_view(['PUT', 'DELETE'])
+def edit_free_board_comment(request, article_pk, comment_pk):
     comment = get_object_or_404(FreeBoardComment, pk=comment_pk, article_id=article_pk)
-    comment.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    # 댓글 수정
+    if request.method == 'PUT':   
+        serializer = FreeBoardCommentSerializer(comment, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    # 댓글 삭제
+    else:
+        comment.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+    
+@api_view(['POST'])
+def like_free_board_article(request, article_pk):
+    article = get_object_or_404(FreeBoardArticle, pk=article_pk)
+    if article.like_user.filter(pk=request.user.pk).exists():
+        # 이미 좋아요를 한 경우, 좋아요 취소
+        article.like_user.remove(request.user)
+        return Response({"message": "게시글 좋아요 취소"}, status=status.HTTP_200_OK)
+    else:
+        # 좋아요 추가
+        article.like_user.add(request.user)
+        return Response({"message": "게시글 좋아요 성공"}, status=status.HTTP_200_OK)
